Administrator.java, Candidate.java, JobOffer.java, Skills.java, User.java
Added class JobOffer with method addToshortlist where a chosen candidate gets addes to a specific shortlist. Added method addcandidate to job offer where you add a Candidate to a list of candidates that apply to a specific list for a job offer.
Added method update Rate List that gets called automatically from method addAcandidate to job offer that calculates the specific rating of a candidate and adds that to a ratelist.Added get Rating method that returns the raterlist position of a candidate.
Added class candidate with method get Type, update Profile,calculate Rating and getters.
Added class administrator that contains a list of all users using the software.It contains a logattempt class called everytime someone tries to enter the system, so as to verificate its username and password.Added class compare 2 candidates that compares two chosen candidates and returns the best of them based on rating. Added method makeAppointment that returns a specific candidate's phoneNumber so as to be able to communicate with company.