Created company class,modified JobOffer class to sort candidates
Created class company,a subclass of User, just like Candidate. Added a constructor and a deleteJobOffer method, that 
could delete an Offer that a company made and would like to get rid of.
Modified class JobOffer, and added a void method named sortCandidates which receives 
a number representing the criterion on which the sorting  is going to be.
0 means that sorting is going to be based on age,1 on degree grade, 2 on
educational level and 3 on work experince.
Added a getOfferName method that retuns the exast name of a job offer.
Added different Sorting classes inside JobOffer using Comparator<Candidate>, so they can be called 
and used based on users liking. If the two users happen to have the same mark on a criterion chosen,
the sorting is based on the candidates rating.