Changed Constructors,Created SortFrame,UpdateProfileFrame,SortFrame
Changed the Constructors in classes Candidate,Company,JobOffer,Skills classes.Added another getter needed in User Class.Created class MainMenuFrame which is the main Frame that the user is going to see the moment he logs in. Created class Update Profile Frame that updates the profile of a user if selected, choosing new Skills. Created class Sort Frame that can be called from a Company, that can sort the enrolled Candidates of a particular JobOffer using a certain criterion.