Created class ApplicationFrame and JobOfferDetails,Modified SortFrame
Created class ApplicationFrame, that shows the Job Offers that a Candidate has applied to.Modified class SortFrame and changed the visualization to match the frame better.Added some error messages when the user does not  give the right input.Created JobOfferDetailsClass that shows in detail job offers from a company. A company can choose to sort the candidates of an offer by clicking on a button.