data, icon, image, Main, MainFrame, RegistryFrame, RegistryFrameCandidate, RegistryFrameCompany, ShortListFrame
Data.ser is the database of the project, Image.png and icon.png are the logo of the project.
Main is the main class of the project, contains the public static variable "f" which is the name of the data file and the public static variable "data" which is the variable that contains the data of the project. The reason that these variables are public : to can be used to all classes.
MainFrame is the main window of the project, where you can log in or sign in, also contains the image.png and some informations at the bottom.
RegistryFrame has only two buttons to choose your type of registry.
RegistryFrameCandidate contains the panel to sign in like a user type of candidate.
RegistryFrameCompany contains the panel to sign in like a user type of company.